requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
python-dotenv
pandas
scikit-learn
numpy
whitenoise
gunicorn